backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.42
psycopg[binary]==3.2.9
PyJWT==2.10.1
pydantic-settings==2.10.1
python-dotenv==1.1.1
argon2-cffi==23.1.0
pytest==8.3.2
httpx==0.28.1
